cfg_getint() callback uses value type 'long', not 'int'

+++ src/config.c
@@ -37,9 +37,9 @@ static int parse_original_envelope(cfg_t* cfg, cfg_opt_t* opt,
                                    const char* value, void* result)
 {
     if (strcasecmp(value, "embedded") == 0)
-        *(int*)result = SRS_ENVELOPE_EMBEDDED;
+        *(long*)result = SRS_ENVELOPE_EMBEDDED;
     else if (strcasecmp(value, "database") == 0)
-        *(int*)result = SRS_ENVELOPE_DATABASE;
+        *(long*)result = SRS_ENVELOPE_DATABASE;
     else
     {
         cfg_error(cfg, "option '%s' must be either 'embedded' or 'database'",
